requirements.txt
mistralai
python-dotenv
streamlit
fuzzywuzzy
openpyxl
scikit-learn